Rename Initialization to UnitTestingModelHelper
